Add Twitch token refresh mechanism + move configs
This commit adds auto refreshing of the Twitch OAuth Token. The token will be validated every 3 minutes and if it needs to be refreshed, it will be refreshed and saved back to the configuration file.
This also moves the configuration files to a volume so that the changes made by the token refresher can persist.
You will also now need to provide a refresh token in the Twitch configuration.